refactor: extract finetune endpoints from orchestrator (#549)
Phase 4.4 of orchestrator decomposition (Strangler Fig):
- Extract 17 LLM finetune endpoints (dataset, training, adapters) to
  modules/llm/router_finetune.py (~260 lines)
- Extract 13 TTS finetune endpoints (samples, training, models) to
  modules/speech/router_finetune.py (~150 lines)
- Move 4 Pydantic models (DatasetProcessRequest, GenerateProjectDatasetRequest,
  FinetuneConfigRequest, AdapterRequest) to LLM finetune router
- Conditional registration: GPU-only (DEPLOYMENT_MODE != "cloud")
- Remove unused imports: get_finetune_manager, get_tts_finetune_manager,
  TTS_FINETUNE_AVAILABLE, File, UploadFile
- Orchestrator: 2875 → 2471 lines (-404)
